restructured current server setup
